+++ tests/_typing/test_error_no_logger.py
@@ -1,0 +1,9 @@
+from hydromt._typing.error import NoDataStrategy, exec_nodata_strat
+
+
+def test_logger_from_frame_in_nodata_strat(caplog):
+    exec_nodata_strat("foo", NoDataStrategy.WARN)
+    assert caplog.records[0].levelname == "WARNING"
+    assert caplog.records[0].message == "foo"
+    # Test that the name of the logger is this current frame's module, not the error.py logger.
+    assert caplog.records[0].name == __name__